pyproject.toml
[project]
name = "agentbeacon"
version = "0.1.0"
description = "AI agent orchestrator"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.6",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "jsonschema>=4.25.1",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
mock-agent = "agentbeacon.mock_agent.__main__:main"

[dependency-groups]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=7.0.0",
    "requests>=2.25.0",
    "psutil>=5.8.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.2",
    "psycopg2-binary>=2.9.0", # PostgreSQL driver for test database cleanup
    "ruff>=0.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentbeacon"]

[tool.pytest.ini_options]
markers = [
    "claude: tests requiring Claude auth (run with pytest -m claude)",
    "copilot: tests requiring Copilot/GitHub auth (run with pytest -m copilot)",
]
addopts = "-m 'not claude and not copilot'"
